name space updated, obsolete code removed
