+++ Git/hm7/CarPark/SearchRequest.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Git.hm7.Transport
+namespace Git.hm7.CarPark
 {
     internal class SearchRequest
     {
