+++ Git/hm7/CarPark/Taxi.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Git.hm7.Transport
+namespace Git.hm7.CarPark
 {
-    internal class Tram : Transports
+    internal class Taxi : Transports
     {
         public override List<TransportType> GetTransportType()
         {
-            return new List<TransportType> { TransportType.Rail, TransportType.Electric };
+            return new List<TransportType> { TransportType.Wheeled, TransportType.Electric };
         }
     }
 }
