+++ Git/hm7/CarPark/Tram.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Git.hm7.Transport
+namespace Git.hm7.CarPark
 {
-    internal class Trolleybus : Transports
+    internal class Tram : Transports
     {
         public override List<TransportType> GetTransportType()
         {
-            return new List<TransportType> { TransportType.Wheeled, TransportType.Electric };
+            return new List<TransportType> { TransportType.Rail, TransportType.Electric };
         }
     }
 }
