+++ Git/hm7/CarPark/TransportType.cs
@@ -4,9 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Git.hm7.Transport
+namespace Git.hm7.CarPark
 {
-    internal class RailTransport
+    public enum TransportType
     {
+        Electric,
+        Rail,
+        Wheeled
     }
 }
