+++ Git/hm7/CarPark/Trolleybus.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Git.hm7.Transport
+namespace Git.hm7.CarPark
 {
-    internal class Taxi : Transports
+    internal class Trolleybus : Transports
     {
         public override List<TransportType> GetTransportType()
         {
