+++ Git/hm7/HomeTask7.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
-using Git.hm7.Transport;
+using Git.hm7.CarPark;
 
 namespace Git.hm7
 {
@@ -68,7 +68,6 @@ namespace Git.hm7
             return from item in transportList
                    where item.Destination == destination && item.DepartureTime == departureTimeUpdated
                    select item;
-            Thread.Sleep(5000);
         }
