--- Git/hm7/Transport/PetrolTransport.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Git.hm7.Transport
-{
-    internal class PetrolTransport
-    {
-    }
-}
