--- Git/hm7/Transport/TransportType.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Git.hm7.Transport
-{
-    public enum TransportType
-    {
-        Electric,
-        Rail,
-        Wheeled
-    }
-}
